Moving `sequence_to_str` to `torchvision._utils`
