+++ torchvision/prototype/utils/_internal.py
@@ -28,9 +28,10 @@ from typing import (
 
 import numpy as np
 import torch
+from torchvision._utils import sequence_to_str
+
 
 __all__ = [
-    "sequence_to_str",
     "add_suggestion",
     "FrozenMapping",
     "make_repr",
@@ -41,18 +42,6 @@ __all__ = [
     "apply_recursively",
     "query_recursively",
 ]
-
-
-def sequence_to_str(seq: Sequence, separate_last: str = "") -> str:
-    if not seq:
-        return ""
-    if len(seq) == 1:
-        return f"'{seq[0]}'"
-
-    head = "'" + "', '".join([str(item) for item in seq[:-1]]) + "'"
-    tail = f"{'' if separate_last and len(seq) == 2 else ','} {separate_last}'{seq[-1]}'"
-
-    return head + tail
 
 
 def add_suggestion(
